+++ schema.sql
@@ -1,0 +1,5 @@
+CREATE TABLE SparkTransformations(
+        ID serial primary key,
+        name varchar,
+        transformation varchar
+);