+++ src/main/java/spark/JeffsProject1Servlet.java
@@ -1,22 +1,27 @@
 package spark;
 
 import java.io.IOException;
+import java.util.LinkedHashMap;
 
 import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-
-@WebServlet(value = "/hello")
-public class HelloServlet extends HttpServlet {
+public class JeffsProject1Servlet extends HttpServlet {
 
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 
 		//retrieve data from database
-		//persistToDatabase().readAll();
+		SQLRepo sqlRepo = new SQLRepo();
+		sqlRepo.readAll();
+		
+		LinkedHashMap<String,String> databaseData = sqlRepo.readAll();
+
+		for(String key2 : databaseData.keySet()) {
+			resp.getWriter().println(databaseData.get(key2));
+		}
 
 	}