+++ src/main/java/spark/serverApp.java
@@ -1,17 +1,26 @@
 package spark;
 
+import java.io.File;
+import java.util.LinkedHashMap;
+
+import org.apache.catalina.LifecycleException;
+import org.apache.catalina.startup.Tomcat;
 import org.apache.spark.SparkConf;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
 
-public class serverApp {
+public class ServerApp {
 	
 	public static void main(String args[]) {
 		
+		//runs server
 		if (args[0].equalsIgnoreCase("server")) {
+			startTomcat();
 			
-			
-		} else if (args[0].equalsIgnoreCase("loadtext")) {
+		} 
+		
+		// used to upload text file data and run server
+		else if (args[0].equalsIgnoreCase("loadtext")) {
 			String textFile = args[1];
 			
 			SparkConf conf = new SparkConf()
@@ -24,44 +33,67 @@ public class serverApp {
 			JavaRDD<MyCSVFile> data = new LoadCSVFile(textFile).createRDD(sc);
 			
 			//spark manipulation
-			SparkTransformation sparkTransformation = new SparkTransformation(data);
+			SparkTransformations sparkTransformations = new SparkTransformations(data);
 			
-			System.out.println(sparkTransformation.countGender().collect());
+			System.out.println(sparkTransformations.getDataStorage("countGender"));
 			
-			System.out.println(sparkTransformation.countRace().collect());
+			System.out.println(sparkTransformations.getDataStorage("countRace"));
 			
-			System.out.println(sparkTransformation.countParentEducation().collect());
+			System.out.println(sparkTransformations.getDataStorage("countParentEducation"));
 			
-			System.out.println(sparkTransformation.countLunch().collect());
+			System.out.println(sparkTransformations.getDataStorage("countLunch"));
 			
-			System.out.println(sparkTransformation.countTestPrep().collect());
+			System.out.println(sparkTransformations.getDataStorage("countTestPrep"));
 			
-			System.out.println("Math Score Avg is: " + sparkTransformation.mathScoreAvg());
+			System.out.println("Math Score Avg is: " + sparkTransformations.getDataStorage("mathScoreAvg"));
 
-			System.out.println("Reading Score Avg is: " + sparkTransformation.readingScoreAvg());
+			System.out.println("Reading Score Avg is: " + sparkTransformations.getDataStorage("readingScoreAvg"));
 
-			System.out.println("Writing Score Avg is: " + sparkTransformation.writingScoreAvg());
+			System.out.println("Writing Score Avg is: " + sparkTransformations.getDataStorage("writingScoreAvg"));
 			
-			System.out.println("Overall Score Avg is: " + sparkTransformation.overallScoreAvg());
+			System.out.println("Overall Score Avg is: " + sparkTransformations.getDataStorage("overallScoreAvg"));
 			
-			System.out.println("Overall Gender Score Avg is: " + sparkTransformation.genderAvg().collect());
+			System.out.println("Overall Gender Score Avg is: " + sparkTransformations.getDataStorage("genderAvg"));
 			
-			System.out.println("Overall Race Score Avg is: " + sparkTransformation.raceAvg().collect());
+			System.out.println("Overall Race Score Avg is: " + sparkTransformations.getDataStorage("raceAvg"));
 			
-			System.out.println("Overall ParentEducation Score Avg is: " + sparkTransformation.parentEducationAvg().collect());
+			System.out.println("Overall ParentEducation Score Avg is: " + sparkTransformations.getDataStorage("parentEducationAvg"));
 			
-			System.out.println("Overall Lunch Score Avg is: " + sparkTransformation.lunchAvg().collect());
+			System.out.println("Overall Lunch Score Avg is: " + sparkTransformations.getDataStorage("lunchAvg"));
 			
-			System.out.println("Overall TestPrep Score Avg is: " + sparkTransformation.testPrepAvg().collect());
+			System.out.println("Overall TestPrep Score Avg is: " + sparkTransformations.getDataStorage("testPrepAvg"));
+			
 			
 			//send to to database
-			//new persistToDatabase().insert(e);
+			LinkedHashMap<String, String> dataStorage = sparkTransformations.getDataStorage();
+			SQLRepo sqlRepo = new SQLRepo();
 			
-			//tomcat pulls from database
-			//call tomcat
+			for(String key : dataStorage.keySet()) {
+				sqlRepo.insert(key,dataStorage.get(key));
+			}
+
+			//initiate embed tomcat
+			startTomcat();
+			
 			//close JavaSparkContect
-			//sc.close();
+			sc.close();
 		}
+	}
+
+	private static void startTomcat() {
+		Tomcat tomcat = new Tomcat();
+		tomcat.setBaseDir(new File("target/tomcat/").getAbsolutePath());
+		tomcat.setPort(8080);
+		tomcat.getConnector();
+		tomcat.addWebapp("/spark", new File("src/main/resources/").getAbsolutePath());
+		tomcat.addServlet("/spark", "JeffsProject1Servlet", new JeffsProject1Servlet()).addMapping("/JeffsProject1");
+		try {
+			tomcat.start();
+		} catch (LifecycleException ex) {
+
+			System.err.println(ex.getMessage());
+		}
+		
 	}
 
 }